fix(odoo): address lead review comments
